refactor: cleanup unused global utils
