+++ packages/blocks/src/_common/utils/string.ts
