+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-shape-button.ts
@@ -5,7 +5,6 @@ import '../panel/shape-panel.js';
 import './change-text-menu.js';
 import './component-toolbar-menu-divider.js';
 
-import { countBy, maxBy } from '@blocksuite/global/utils';
 import { WithDisposable } from '@blocksuite/lit';
 import type { Page } from '@blocksuite/store';
 import { css, html, LitElement, nothing } from 'lit';
@@ -19,6 +18,7 @@ import {
   ShapeArrowDownSmallIcon,
 } from '../../../../_common/icons/index.js';
 import type { CssVariableName } from '../../../../_common/theme/css-variables.js';
+import { countBy, maxBy } from '../../../../_common/utils/iterable.js';
 import { LineWidth, type ShapeTool } from '../../../../_common/utils/types.js';
 import type { PhasorElementType } from '../../../../surface-block/index.js';
 import {
