--- packages/global/src/error/code.ts
@@ -1,4 +1,0 @@
-export enum ErrorCode {
-  MIGRATION_ERROR = 100,
-  SCHEMA_VALIDATE_ERROR,
-}
