--- packages/global/src/error/index.ts
@@ -1,3 +1,0 @@
-export * from './code.js';
-export * from './error.js';
-export * from './guard.js';
